views,usage: Move query specs to their own file
