+++ usage/client.go
@@ -3,7 +3,6 @@ package usage
 import (
 	"context"
 	"fmt"
-	"strings"
 
 	"cloud.google.com/go/bigquery"
 	livepeer "github.com/livepeer/go-api-client"
@@ -79,24 +78,6 @@ func toFloat64Ptr(bqFloat bigquery.NullFloat64, asked bool) data.Nullable[float6
 
 func toStringPtr(bqStr bigquery.NullString, asked bool) data.Nullable[string] {
 	return data.ToNullable(bqStr.StringVal, bqStr.Valid, asked)
-}
-
-func (q *QuerySpec) HasAnyBreakdown() bool {
-	return len(q.BreakdownBy) > 0 || q.TimeStep != ""
-}
-
-func (q *QuerySpec) hasBreakdownBy(e string) bool {
-	// callers always set `e` as a string literal so we can panic if it's not valid
-	if usageBreakdownFields[e] == "" {
-		panic(fmt.Sprintf("unknown breakdown field %q", e))
-	}
-
-	for _, a := range q.BreakdownBy {
-		if strings.EqualFold(a, e) {
-			return true
-		}
-	}
-	return false
 }
 
 func (c *Client) QuerySummaryWithBreakdown(ctx context.Context, spec QuerySpec) ([]Metric, error) {
